+++ receiver/hostmetricsreceiver/internal/scraper/diskscraper/disk_scraper_windows_test.go
@@ -25,8 +25,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal/perfcounters"
+	"go.opentelemetry.io/collector/receiver/scrapererror"
 )
 
 func TestScrape_Error(t *testing.T) {
@@ -69,10 +69,10 @@ func TestScrape_Error(t *testing.T) {
 			_, err = scraper.scrape(context.Background())
 			assert.EqualError(t, err, test.expectedErr)
 
-			isPartial := consumererror.IsPartialScrapeError(err)
+			isPartial := scrapererror.IsPartialScrapeError(err)
 			assert.True(t, isPartial)
 			if isPartial {
-				assert.Equal(t, metricsLen, err.(consumererror.PartialScrapeError).Failed)
+				assert.Equal(t, metricsLen, err.(scrapererror.PartialScrapeError).Failed)
 			}
 		})
 	}