+++ receiver/hostmetricsreceiver/internal/scraper/processscraper/process_scraper_test.go
@@ -28,11 +28,11 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/internal/processor/filterset"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal/metadata"
+	"go.opentelemetry.io/collector/receiver/scrapererror"
 	"go.opentelemetry.io/collector/translator/conventions"
 )
 
@@ -62,9 +62,9 @@ func TestScrape(t *testing.T) {
 	//
 	// so validate that we have at some errors & some valid data
 	if err != nil {
-		require.True(t, consumererror.IsPartialScrapeError(err))
+		require.True(t, scrapererror.IsPartialScrapeError(err))
 		noProcessesScraped := resourceMetrics.Len()
-		noProcessesErrored := err.(consumererror.PartialScrapeError).Failed
+		noProcessesErrored := err.(scrapererror.PartialScrapeError).Failed
 		require.Lessf(t, 0, noProcessesErrored, "Failed to scrape metrics - : error, but 0 failed process %v", err)
 		require.Lessf(t, 0, noProcessesScraped, "Failed to scrape metrics - : 0 successful scrapes %v", err)
 	}
@@ -174,7 +174,7 @@ func TestScrapeMetrics_GetProcessesError(t *testing.T) {
 	metrics, err := scraper.scrape(context.Background())
 	assert.EqualError(t, err, "err1")
 	assert.Equal(t, 0, metrics.Len())
-	assert.False(t, consumererror.IsPartialScrapeError(err))
+	assert.False(t, scrapererror.IsPartialScrapeError(err))
 }
 
 type processHandlesMock struct {
@@ -447,11 +447,11 @@ func TestScrapeMetrics_ProcessErrors(t *testing.T) {
 			assert.Equal(t, expectedMetricsLen, md.MetricCount())
 
 			assert.EqualError(t, err, test.expectedError)
-			isPartial := consumererror.IsPartialScrapeError(err)
+			isPartial := scrapererror.IsPartialScrapeError(err)
 			assert.True(t, isPartial)
 			if isPartial {
 				expectedFailures := getExpectedScrapeFailures(test.nameError, test.exeError, test.timesError, test.memoryInfoError, test.ioCountersError)
-				assert.Equal(t, expectedFailures, err.(consumererror.PartialScrapeError).Failed)
+				assert.Equal(t, expectedFailures, err.(scrapererror.PartialScrapeError).Failed)
 			}
 		})
 	}